Reorganiza estructura de archivos
