+++ packages/reciclateya-theme/src/components/Home.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { connect, styled, css } from "frontity";
-import Link from "./components/Link";
-import { renderText } from "./utils/helpers";
+import Link from "./Link";
+import { renderText } from "../utils/helpers";
 
 const Home = ({ state, posts }) => {
   const Items = styled.div`
